# specs/master/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Azure OpenAI Local Middleware
  description: |
    Local FastAPI proxy for Azure OpenAI with authentication, logging, and cost tracking.
    Maintains full compatibility with Azure OpenAI REST API.
  version: 1.0.0
  contact:
    name: Azure Middleware
servers:
  - url: http://localhost:8000
    description: Local development server
security:
  - localApiKey: []
paths:
...
  /openai/deployments/{deployment}/embeddings:
    post:
      summary: Embeddings
      description: Create embeddings for input text. Fully compatible with Azure OpenAI API.
      operationId: createEmbeddings
      parameters:
        - $ref: '#/components/parameters/deployment'
        - $ref: '#/components/parameters/apiVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingRequest'
      responses:
        '200':
          description: Embedding response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/CostCapExceeded'
        '502':
          $ref: '#/components/responses/BadGateway'
  /openai/deployments/{deployment}/responses:
    post:
      summary: Responses API
      description: Create a response using the Responses API. Fully compatible with Azure OpenAI API.
      operationId: createResponse
      parameters:
        - $ref: '#/components/parameters/deployment'
        - $ref: '#/components/parameters/apiVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Responses API request body (passthrough)
      responses:
        '200':
          description: Response API response
          content:
            application/json:
              schema:
                type: object
                description: Responses API response (passthrough)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/CostCapExceeded'
        '502':
          $ref: '#/components/responses/BadGateway'
components:
  securitySchemes:
    localApiKey:
      type: apiKey
      in: header
      name: api-key
      description: Local API key for request authentication
  parameters:
    deployment:
      name: deployment
      in: path
      required: true
      description: Azure OpenAI deployment name
      schema:
        type: string
        example: gpt-4
    apiVersion:
      name: api-version
      in: query
      required: false
      description: Azure OpenAI API version (defaults to config value)
      schema:
        type: string
        example: "2024-02-01"
  schemas:
...
    EmbeddingRequest:
      type: object
      required:
        - input
      properties:
        input:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        encoding_format:
          type: string
          enum: [float, base64]
          default: float
    EmbeddingResponse:
      type: object
      required:
        - object
        - data
        - model
        - usage
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: '#/components/schemas/Embedding'
        model:
          type: string
        usage:
          $ref: '#/components/schemas/EmbeddingUsage'
    Embedding:
      type: object
      required:
        - object
        - embedding
        - index
      properties:
        object:
          type: string
          enum: [embedding]
        embedding:
          type: array
          items:
            type: number
        index:
          type: integer
...
    EmbeddingUsage:
      type: object
      required:
        - prompt_tokens
        - total_tokens
      properties:
        prompt_tokens:
          type: integer
        total_tokens:
          type: integer
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            param:
              type: string
            type:
              type: string
    CostCapError:
      type: object
      required:
        - error
        - current_cost_eur
        - limit_eur
        - message
      properties:
        error:
          type: string
          enum: [daily_cost_limit_exceeded]
        current_cost_eur:
          type: number
          format: float
        limit_eur:
          type: number
          format: float
        message:
          type: string
        retry_after:
          type: integer
          description: Seconds until UTC midnight
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    CostCapExceeded:
      description: Daily cost cap exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until UTC midnight
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CostCapError'
    BadGateway:
      description: Azure OpenAI unreachable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'